A change in print statement
This program was calculating the area in every case and printing it after the control is out of switch case, but what if the user enters a wrong choice and in that case the default case has to be executed then no area is calculated there. When the control will be out of switch and the next statement to executed is printf and there will be no value in the variable area. Hence it will display a garbage value. I have made some changes to resolve the same.